+++ frontend/app/src/main/java/com/example/shareader/MainActivity.kt
@@ -36,66 +36,3 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
-@Composable
-fun HomeView() {
-    val context = LocalContext.current
-    val activityLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) { _ ->
-    }
-
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colorScheme.background
-    ) {
-        Column (
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-            modifier = Modifier.padding()
-        ){
-            Button(
-                onClick = {
-                    val intent = Intent(context, ReaderActivity::class.java)
-                    activityLauncher.launch(intent)
-                },
-                modifier = Modifier
-                    .width(200.dp)
-            ) {
-                Text(
-                    text = "E-Book Reader",
-                    color = Color.White,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-            Button(
-                onClick = {
-                    val intent = Intent(context, SummaryActivity::class.java)
-                    activityLauncher.launch(intent)
-                },
-                modifier = Modifier
-                    .width(200.dp)
-            ) {
-                Text(
-                    text = "Summary View",
-                    color = Color.White,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-            Button(
-                onClick = {
-                    val intent = Intent(context, QuizActivity::class.java)
-                    activityLauncher.launch(intent)
-                },
-                modifier = Modifier
-                    .width(200.dp)
-            ) {
-                Text(
-                    text = "Quiz View",
-                    color = Color.White,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-        }
-    }
-}
